Move code defining Formulation and Catchment_Formulation to standalone cpp files
